requirements.txt
streamlit
langchain
langchain-community
langchain-core
openai
chromadb
tiktoken
python-dotenv
faiss-cpu
Pillow
gspread
google-auth
